split the components, got it working
